+++ sources/Google.Solutions.IapDesktop.Core.Test/Google.Solutions.IapDesktop.Core.Test.csproj
@@ -90,9 +90,9 @@
     <Compile Include="ClientModel\Protocol\TestAppProtocolComand.cs" />
     <Compile Include="ClientModel\Protocol\TestAppProtocolFactory.cs" />
     <Compile Include="ClientModel\Protocol\TestAppProtocolFactoryConfigurationSection.cs" />
-    <Compile Include="ClientModel\Traits\TestLinuxTrait.cs" />
-    <Compile Include="ClientModel\Traits\TestWindowsTrait.cs" />
-    <Compile Include="ClientModel\Traits\TestInstanceTrait.cs" />
+    <Compile Include="ProjectModel\Traits\TestLinuxTrait.cs" />
+    <Compile Include="ProjectModel\Traits\TestWindowsTrait.cs" />
+    <Compile Include="ProjectModel\Traits\TestInstanceTrait.cs" />
     <Compile Include="ClientModel\Transport\Policies\TestCurrentProcessPolicy.cs" />
     <Compile Include="ClientModel\Transport\Policies\TestAllowAllPolicy.cs" />
     <Compile Include="ClientModel\Transport\TestDirectTransportFactory.cs" />
