+++ sources/Google.Solutions.IapDesktop.Core.Test/ProjectModel/Traits/TestWindowsTrait.cs
@@ -21,17 +21,18 @@
 
 using Google.Solutions.IapDesktop.Core.ClientModel.Protocol;
 using Google.Solutions.IapDesktop.Core.ClientModel.Traits;
+using Google.Solutions.IapDesktop.Core.ProjectModel.Traits;
 using Google.Solutions.Testing.Apis;
 using NUnit.Framework;
 
-namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Traits
+namespace Google.Solutions.IapDesktop.Core.Test.ProjectModel.Traits
 {
     [TestFixture]
-    public class TestLinuxTrait : EquatableFixtureBase<LinuxTrait, IProtocolTargetTrait>
+    public class TestWindowsTrait : EquatableFixtureBase<WindowsTrait, IProtocolTargetTrait>
     {
-        protected override LinuxTrait CreateInstance()
+        protected override WindowsTrait CreateInstance()
         {
-            return new LinuxTrait();
+            return new WindowsTrait();
         }
 
         //---------------------------------------------------------------------
@@ -41,7 +42,7 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Traits
         [Test]
         public void DisplayNameExpression()
         {
-            Assert.AreEqual("isLinux()", new InstanceTrait().DisplayName);
+            Assert.AreEqual("isWindows()", new InstanceTrait().DisplayName);
         }
 
         //---------------------------------------------------------------------
@@ -51,7 +52,7 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Traits
         [Test]
         public void ToStringReturnsExpression()
         {
-            Assert.AreEqual("isLinux()", new LinuxTrait().ToString());
+            Assert.AreEqual("isWindows()", new WindowsTrait().ToString());
         }
 
         //---------------------------------------------------------------------
@@ -62,14 +63,14 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Traits
         public void WhenExpressionIsNullOrEmpty_ThenTryParseReturnsFalse(
             [Values(" \t", "", null)] string expression)
         {
-            Assert.IsFalse(LinuxTrait.TryParse(expression, out var _));
+            Assert.IsFalse(WindowsTrait.TryParse(expression, out var _));
         }
 
         [Test]
         public void WhenExpressionIsValid_ThenTryParseReturnsTrue(
-            [Values("isLinux()", " isLinux(  \n) \n\r\t ")] string expression)
+            [Values("isWindows()", " isWindows(  \n) \n\r\t ")] string expression)
         {
-            Assert.IsTrue(LinuxTrait.TryParse(expression, out var trait));
+            Assert.IsTrue(WindowsTrait.TryParse(expression, out var trait));
             Assert.IsNotNull(trait);
         }
