+++ sources/Google.Solutions.IapDesktop.Core/Google.Solutions.IapDesktop.Core.csproj
@@ -105,7 +105,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ClientModel\Protocol\IProtocol.cs" />
     <Compile Include="ClientModel\Protocol\IProtocolTarget.cs" />
-    <Content Include="ClientModel\Traits\TypeTrait.tt">
+    <Content Include="ProjectModel\Traits\TypeTrait.tt">
       <Generator>TextTemplatingFileGenerator</Generator>
       <LastGenOutput>TypeTrait.cs</LastGenOutput>
     </Content>
