+++ sources/Google.Solutions.IapDesktop.Core/ProjectModel/Traits/TypeTrait.tt
@@ -27,7 +27,7 @@
 using Google.Solutions.IapDesktop.Core.ClientModel.Protocol;
 using System.Linq;
 
-namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
+namespace Google.Solutions.IapDesktop.Core.ProjectModel.Traits
 {
 	<#
 	var traitNames = new [] {"Instance", "Windows", "Linux"};
